@@ -6,13 +7,16 @@
 	<version>1.0.0</version>
 
 	<properties>
-		<java.version>1.8</java.version>
+		<java.version>1.9</java.version>
+		<springboot.version>2.0.5.RELEASE</springboot.version>
+		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+		<project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
 	</properties>
 
 	<parent>
 		<groupId>org.springframework.boot</groupId>
 		<artifactId>spring-boot-starter-parent</artifactId>
-		<version>1.5.8.RELEASE</version>
+		<version>2.0.5.RELEASE</version>
 	</parent>
 
 	<dependencies>
